+++ core/objects/finders/Bvh.h
@@ -126,6 +126,8 @@ private:
     Float rSqr;
 
 public:
+    BvhSphere() = default;
+
     BvhSphere(const Vector& center, const Float radius)
         : center(center)
         , r(radius)